fix #1 Start the project structure with Maven... This commit fixes the issue that the jar file was not assembled correctly. Run this command: "mvn clean compile assembly:single" to create a single jar file with all dependencies. The server can then be started by running "java -cp target/CI-1.0-SNAPSHOT-jar-with-dependencies.jar group1.Main" This will make further development easier, by making possible automatic dependency management possible. Just add the missing dependency in the  pom.xml
