+++ CI/src/main/java/group1/Main.java
@@ -1,0 +1,22 @@
+package group1;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import javax.servlet.ServletException;
+ 
+import java.io.IOException;
+ 
+import org.eclipse.jetty.server.Server;
+import org.eclipse.jetty.server.Request;
+import org.eclipse.jetty.server.handler.AbstractHandler;
+class Main
+{
+    // used to start the CI server in command line
+    public static void main(String[] args) throws Exception
+    {
+        Server server = new Server(8080);
+        server.setHandler(new ContinuousIntegrationServer()); 
+        server.start();
+        server.join();
+    }
+}
